ban is now more in sync with mute, merged trackerstatus files into one
