+++ chiya/cogs/mute.py
@@ -2,7 +2,6 @@ import arrow
 import discord
 from discord import app_commands
 from discord.ext import commands
-from loguru import logger
 from parsedatetime import Calendar
 
 from chiya.config import config
